Create a Gauss-Markov (GM) Process

Sets up the necessary backend for the GM process.

GM(beta = NULL, sigma2_gm = 1)

Arguments

beta
A double value for the \(\beta\) of an GM process (see Note for details).

sigma2_gm
A double value for the variance, \(\sigma ^2_{gm}\), of a GM process (see Note for details).

Value

An S3 object with called ts.model with the following structure:

process.desc

Used in summary: "BETA","SIGMA2"

theta

\(\beta\), \(\sigma ^2_{gm}\)

plength

Number of parameters

print

String containing simplified model

desc

"GM"

obj.desc

Depth of parameters e.g. list(1,1)

starting

Guess starting values? TRUE or FALSE (e.g. specified value)

Details

When supplying values for \(\beta\) and \(\sigma ^2_{gm}\), these parameters should be of a GM process and NOT of an AR1. That is, do not supply AR1 parameters such as \(\phi\), \(\sigma^2\).

Internally, GM parameters are converted to AR1 using the `freq` supplied when creating data objects (gts) or specifying a `freq` parameter in simts or simts.imu.

The `freq` of a data object takes precedence over the `freq` set when modeling.

Note

We consider the following model: $$X_t = e^{(-\beta)} X_{t-1} + \varepsilon_t$$, where \(\varepsilon_t\) is iid from a zero mean normal distribution with variance \(\sigma^2(1-e^{2\beta})\).

Author

James Balamuta

Examples

GM()
GM(beta=.32, sigma2_gm=1.3)