Create a Seasonal Autoregressive Integrated Moving Average (SARIMA) Process

Sets up the necessary backend for the SARIMA process.

SARIMA(ar = 1, i = 0, ma = 1, sar = 1, si = 0, sma = 1,
  s = 12, sigma2 = 1)

Arguments

ar
A vector or integer containing either the coefficients for \(\phi\)'s or the process number \(p\) for the Autoregressive (AR) term.

i
An integer containing the number of differences to be done.

ma
A vector or integer containing either the coefficients for \(\theta\)'s or the process number \(q\) for the Moving Average (MA) term.

sar
A vector or integer containing either the coefficients for \(\Phi\)'s or the process number \(P\) for the Seasonal Autoregressive (SAR) term.

si
An integer containing the number of seasonal differences to be done.

sma
A vector or integer containing either the coefficients for \(\Theta\)'s or the process number \(Q\) for the Seasonal Moving Average (SMA) term.

s
An integer containing the seasonality.

sigma2
A double value for the standard deviation, \(\sigma\), of the SARMA process.

Value

An S3 object with called ts.model with the following structure:

process.desc

\(AR*p\), \(MA*q\), \(SAR*P\), \(SMA*Q\)

theta

\(\sigma\)

plength

Number of parameters

desc

Type of model

desc.simple

Type of model (after simplification)

print

String containing simplified model

obj.desc

y desc replicated x times

obj

Depth of Parameters e.g. list(c(length(ar), length(ma), length(sar), length(sma), 1, i, si) )

starting

Guess Starting values? TRUE or FALSE (e.g. specified value)

Details

A variance is required since the model generation statements utilize randomization functions expecting a variance instead of a standard deviation unlike R.

Examples

# Create an SARIMA(1,1,2)x(1,0,1) process
SARIMA(ar = 1, i = 1, ma = 2, sar = 1, si = 0, sma =1)

# Creates an SARMA(1,0,1)x(1,1,1) process with predefined coefficients.
SARIMA(ar=0.23, i = 0, ma=0.4, sar = .3,  sma = .3)