ARMA process to WV

This function computes the Haar Wavelet Variance of an ARMA process

arma_to_wv(ar, ma, sigma2, tau)

Arguments

ar
A vec containing the coefficients of the AR process

ma
A vec containing the coefficients of the MA process

sigma2
A double containing the residual variance

tau
A vec containing the scales e.g. \(2^{\tau}\)

Value

A vec containing the wavelet variance of the ARMA process.

Details

The function is a generic implementation that requires a stationary theoretical autocorrelation function (ACF) and the ability to transform an ARMA(\(p\),\(q\)) process into an MA(\(\infty\)) (e.g. infinite MA process).

Process Haar Wavelet Variance Formula

The Autoregressive Order \(p\) and Moving Average Order \(q\) (ARMA(\(p\),\(q\))) process has a Haar Wavelet Variance given by: $$\frac{{{\tau _j}\left[ {1 - \rho \left( {\frac{{{\tau _j}}}{2}} \right)} \right] + 2\sum\limits_{i = 1}^{\frac{{{\tau _j}}}{2} - 1} {i\left[ {2\rho \left( {\frac{{{\tau _j}}}{2} - i} \right) - \rho \left( i \right) - \rho \left( {{\tau _j} - i} \right)} \right]} }}{{\tau _j^2}}\sigma _X^2$$ where \(\sigma _X^2\) is given by the variance of the ARMA process. Furthermore, this assumes that stationarity has been achieved as it directly