Time Series Convolution Filters

Applies a convolution filter to a univariate time series.

cfilter(x, filter, sides, circular)

Arguments

x
A column vector of length T

filter
A column vector of length f

sides
An int that takes either 1:for using past values only or 2: filter coefficients are centered around lag 0.

circular
A bool that indicates if the filter should be wrapped around the ends of the time series.

Value

A column vec that contains the results of the filtering process.

Details

This is a port of the cfilter function harnessed by the filter function in stats. It is about 5-7 times faster than R's base function. The benchmark was done on iMac Late 2013 using vecLib as the BLAS.

Examples

x = 1:15
# 
cfilter(x, rep(1, 3), sides = 2, circular = FALSE)

#>       [,1]
#>  [1,]   NA
#>  [2,]    6
#>  [3,]    9
#>  [4,]   12
#>  [5,]   15
#>  [6,]   18
#>  [7,]   21
#>  [8,]   24
#>  [9,]   27
#> [10,]   30
#> [11,]   33
#> [12,]   36
#> [13,]   39
#> [14,]   42
#> [15,]   NA

# Using R's function
filter(x, rep(1, 3))

#> Time Series:
#> Start = 1 
#> End = 15 
#> Frequency = 1 
#>  [1] NA  6  9 12 15 18 21 24 27 30 33 36 39 42 NA

#
cfilter(x, rep(1, 3), sides = 1, circular = FALSE)

#>       [,1]
#>  [1,]   NA
#>  [2,]   NA
#>  [3,]    6
#>  [4,]    9
#>  [5,]   12
#>  [6,]   15
#>  [7,]   18
#>  [8,]   21
#>  [9,]   24
#> [10,]   27
#> [11,]   30
#> [12,]   33
#> [13,]   36
#> [14,]   39
#> [15,]   42

# Using R's function
filter(x, rep(1, 3), sides = 1)

#> Time Series:
#> Start = 1 
#> End = 15 
#> Frequency = 1 
#>  [1] NA NA  6  9 12 15 18 21 24 27 30 33 36 39 42

#
cfilter(x, rep(1, 3), sides = 1, circular = TRUE)

#>       [,1]
#>  [1,]   30
#>  [2,]   18
#>  [3,]    6
#>  [4,]    9
#>  [5,]   12
#>  [6,]   15
#>  [7,]   18
#>  [8,]   21
#>  [9,]   24
#> [10,]   27
#> [11,]   30
#> [12,]   33
#> [13,]   36
#> [14,]   39
#> [15,]   42

# Using R's function
filter(x, rep(1, 3), sides = 1, circular = TRUE)

#> Time Series:
#> Start = 1 
#> End = 15 
#> Frequency = 1 
#>  [1] 30 18  6  9 12 15 18 21 24 27 30 33 36 39 42