Computes the (MODWT) wavelet covariance matrix

Calculates the (MODWT) wavelet covariance matrix

compute_cov_cpp(
  signal_modwt,
  nb_level,
  compute_v = "diag",
  robust = TRUE,
  eff = 0.6
)

Arguments

signal_modwt
A field<vec> that contains the modwt decomposition.

nb_level
A integer that contains the level of decomposition J.

compute_v
A string that indicates what kind of matrix should be created. Possible options: "diag" or "none"

robust
A boolean that triggers the use of the robust estimate.

eff
A double that indicates the efficiency as it relates to an MLE.

Value

A field<mat> containing the covariance matrix.