Internal IMU Object Construction

Internal quick build for imu object.

create_imu(data, ngyros, nacces, axis, freq, unit = NULL, name = NULL,
  stype = NULL)

Arguments

data
A matrix with dimensions N x length(index)

ngyros
An integer containing the number of gyroscopes

axis
A vector unique representation of elements e.g. x,y,z or x,y or x.

freq
An integer that provides the frequency for the data.

unit
A string that contains the unit expression of the frequency. Default value is NULL.

name
A string that provides an identifier to the data. Default value is NULL.

stype
A string that describes the sensor type. Default value is NULL.

naccess
An integer containing the number of accelerometers

Value

An imu object class.