Analytic D matrix for ARMA(1,1) process

Obtain the first derivative of the ARMA(1,1) process.

deriv_arma11(phi, theta, sigma2, tau)

Arguments

phi

A double corresponding to the phi coefficient of an ARMA(1,1) process.

theta

A double corresponding to the theta coefficient of an ARMA(1,1) process.

sigma2

A double corresponding to the error term of an ARMA(1,1) process.

tau

A vec containing the scales e.g. \(2^{\tau}\)

Value

A matrix with:

The first column containing the partial derivative with respect to \(\phi\);

The second column containing the partial derivative with respect to \(\theta\);

The third column contains the partial derivative with respect to \(\sigma ^2\).

Process Haar WV First Derivative

Taking the derivative with respect to \(\phi\) yields: $$ \frac{\partial }{{\partial \phi }}\nu _j^2\left( {\phi ,\theta ,{\sigma ^2}} \right) = \frac{{2{\sigma ^2}}}{{{{(\phi - 1)}^4}{{(\phi + 1)}^2}\tau _j^2}}\left( \begin{array}{cc} &{\tau _j}\left( { - {{(\theta + 1)}^2}(\phi - 1){{(\phi + 1)}^2} - 2\left( {{\phi ^2} - 1} \right)(\theta + \phi )(\theta \phi + 1){\phi ^{\frac{{{\tau _j}}}{2} - 1}} + \left( {{\phi ^2} - 1} \right)(\theta \phi + 1)(\theta + \phi ){\phi ^{{\tau _j} - 1}}} \right) \\ &- \left( {{\theta ^2}((3\phi + 2)\phi + 1) + 2\theta \left( {\left( {{\phi ^2} + \phi + 3} \right)\phi + 1} \right) + (3\phi + 2)\phi + 1} \right)\left( {{\phi ^{{\tau _j}}} - 4{\phi ^{\frac{{{\tau _j}}}{2}}} + 3} \right) \\ \end{array} \right)$$

Taking the derivative with respect to \(\theta\) yields: $$\frac{\partial }{{\partial \theta }}\nu _j^2\left( {\phi ,\theta ,{\sigma ^2}} \right) = \frac{{2{\sigma ^2}\left( {(\theta + 1)\left( {{\phi ^2} - 1} \right){\tau _j} + \left( {2\theta \phi + {\phi ^2} + 1} \right)\left( {{\phi ^{{\tau _j}}} - 4{\phi ^{\frac{{{\tau _j}}}{2}}} + 3} \right)} \right)}}{{{{(\phi - 1)}^3}(\phi + 1)\tau _j^2}}$$

Taking the derivative with respect to \(\sigma^2\) yields: $$\frac{\partial }{{\partial \sigma ^2 }}\nu _j^2\left( {\phi ,\theta ,{\sigma ^2}} \right) = \frac{2 \sigma ^2 \left(\left(\phi ^2-1\right) \tau _j+2 \phi \left(\phi ^{\tau _j}-4 \phi ^{\frac{\tau _j}{2}}+3\right)\right)}{(\phi -1)^3 (\phi +1) \tau _j^2}$$

Author

James Joseph Balamuta (JJB)