Evalute a time series or a list of time series models

This function calculates AIC, BIC and HQ or the MAPE for a list of time series models. This function currently only supports models estimated by the MLE.

evaluate(
  models,
  Xt,
  criterion = "IC",
  start = 0.8,
  demean = TRUE,
  print = TRUE
)

Arguments

models

A time series model or a list of time series models.

Xt

A time series (i.e gts object).

criterion

Either "IC" for AIC, BIC and HQ or "MAPE" for MAPE.

start

A numeric indicating the starting proportion of the data that is used for prediction (assuming criterion = "MAPE").

demean

A boolean indicating whether the model includes a mean / intercept term or not.

print

logical. If TRUE (the default) results are printed.

Value

AIC, BIC and HQ or MAPE

Author

Stéphane Guerrier

Examples

set.seed(18)
n = 300
Xt = gen_gts(n, AR(phi = c(0, 0, 0.8), sigma2 = 1))
evaluate(AR(1), Xt)
#>            AIC      BIC       HQ
#> AR(1) 1105.536 1116.647 1109.982
#> 
evaluate(list(AR(1), AR(3), MA(3), ARMA(1,2), 
SARIMA(ar = 1, i = 0, ma = 1, sar = 1, si = 1, sma = 1, s = 12)), Xt)
#>                                    AIC       BIC        HQ
#> AR(1)                        1105.5356 1116.6469 1109.9824
#> AR(3)                         863.6250  882.1439  871.0363
#> MA(3)                         959.1555  977.6744  966.5668
#> ARMA(1,2)                    1015.2700 1033.7889 1022.6813
#> SARIMA(1,0,1) x (1,1,1) [12] 1075.4877 1094.0066 1082.8990
#> 
#> AIC suggests: AR(3)
#> BIC suggests: AR(3)
#> HQ suggests : AR(3)
evaluate(list(AR(1), AR(3)), Xt, criterion = "MAPE")
#>            MAPE        SD
#> AR(1) 1.0416358 0.1638334
#> AR(3) 0.6846872 0.1119464
#> 
#> MAPE suggests: AR(3)