Generate Latent Time Series based on Model (Internal)

Create a latent time series based on a supplied time series model.

gen_lts_cpp(N, theta, desc, objdesc)

Arguments

N
An interger containing the amount of observations for the time series.

theta
A vec containing the parameters to use to generate the model.

desc
A vector<string> containing the different model types (AR1, WN, etc..).

objdesc
A field<vec> containing the different model objects e.g. AR1 = c(1,1)

Value

A mat containing data for each decomposed and combined time series.

Examples

# AR
set.seed(1336)
gen_lts_cpp(10, c(.9,1), "AR1", list(c(1,1)))

#>             [,1]       [,2]
#>  [1,] -1.2646786 -1.2646786
#>  [2,] -1.1335137 -1.1335137
#>  [3,] -0.6978944 -0.6978944
#>  [4,] -0.9467181 -0.9467181
#>  [5,] -2.3142463 -2.3142463
#>  [6,] -2.6023244 -2.6023244
#>  [7,] -2.7281407 -2.7281407
#>  [8,] -0.7326329 -0.7326329
#>  [9,]  0.4918010  0.4918010
#> [10,] -0.9250607 -0.9250607