Generate Time Series based on Model (Internal)

Create a time series process based on a supplied ts.model.

gen_model(N, theta, desc, objdesc)

Arguments

N
An interger containing the amount of observations for the time series.

theta
A vec containing the parameters to use to generate the model

desc
A vector<string> containing the different model types (AR1, WN, etc..)

objdesc
A field<vec> contains the different model objects e.g. AR1 = c(1,1)

Value

A vec that contains combined time series.

Examples

# AR
set.seed(1336)
gen_model(1000, c(.9,1), "AR1", list(c(1,1)))

#>                 [,1]
#>    [1,] -1.264678619
#>    [2,] -1.133513748
#>    [3,] -0.697894433
#>    [4,] -0.946718140
#>    [5,] -2.314246265
#>    [6,] -2.602324387
#>    [7,] -2.728140748
#>    [8,] -0.732632909
#>    [9,]  0.491800967
#>   [10,] -0.925060720
#>   [11,] -1.383921971
#>   [12,] -0.640876120
#>   [13,]  0.137320557
#>   [14,]  1.625811149
#>   [15,]  1.659588424
#>   [16,]  1.242602081
#>   [17,]  1.144019324
#>   [18,]  1.499085819
#>   [19,]  2.988645823
#>   [20,]  4.811756186
#>   [21,]  4.203589906
#>   [22,]  2.034724209
#>   [23,]  4.477198246
#>   [24,]  4.288238790
#>   [25,]  4.461805953
#>   [26,]  5.358950042
#>   [27,]  4.539133049
#>   [28,]  4.699648774
#>   [29,]  5.309557385
#>   [30,]  7.526941093
#>   [31,]  6.787245021
#>   [32,]  6.107997036
#>   [33,]  4.676997545
#>   [34,]  4.234361777
#>   [35,]  4.339251151
#>   [36,]  3.267808500
#>   [37,]  4.448306250
#>   [38,]  3.039984571
#>   [39,]  1.937410277
#>   [40,]  1.086082141
#>   [41,]  0.193687477
#>   [42,] -1.171424362
#>   [43,] -1.583765371
#>   [44,] -1.587208970
#>   [45,] -1.351502366
#>   [46,] -2.512288782
#>   [47,] -3.570327468
#>   [48,] -4.881864268
#>   [49,] -3.433185536
#>   [50,] -2.863959083
#>   [51,] -1.575293875
#>   [52,] -2.421002444
#>   [53,] -3.290149646
#>   [54,] -2.270029806
#>   [55,] -1.172474387
#>   [56,] -2.265697948
#>   [57,] -1.810259986
#>   [58,] -1.335367226
#>   [59,] -0.799625835
#>   [60,] -1.401105943
#>   [61,] -4.346476135
#>   [62,] -3.591701368
#>   [63,] -4.992692162
#>   [64,] -5.653233332
#>   [65,] -4.992994676
#>   [66,] -4.724355640
#>   [67,] -5.476044395
#>   [68,] -5.357557877
#>   [69,] -4.964034277
#>   [70,] -3.994641995
#>   [71,] -3.382237021
#>   [72,] -3.117430111
#>   [73,] -3.114303620
#>   [74,] -1.548196500
#>   [75,] -1.105129167
#>   [76,]  0.842280700
#>   [77,] -0.245584463
#>   [78,] -0.785450759
#>   [79,] -0.842270946
#>   [80,] -1.010011179
#>   [81,] -1.153467840
#>   [82,] -1.067373805
#>   [83,] -1.375288156
#>   [84,] -1.286650726
#>   [85,] -1.743805221
#>   [86,] -1.770356278
#>   [87,] -1.881552015
#>   [88,] -2.161519559
#>   [89,] -2.322426601
#>   [90,] -2.079235506
#>   [91,] -2.106952860
#>   [92,]  0.209886923
#>   [93,] -0.223885263
#>   [94,]  0.116803010
#>   [95,]  0.384350265
#>   [96,]  0.779173154
#>   [97,]  1.684025318
#>   [98,]  2.946232460
#>   [99,]  2.472264909
#>  [100,]  3.399965061
#>  [101,]  3.156641844
#>  [102,]  3.739990360
#>  [103,]  2.299274010
#>  [104,]  1.547635100
#>  [105,]  1.184549941
#>  [106,] -0.128313827
#>  [107,] -0.622566589
#>  [108,]  0.282632201
#>  [109,]  0.902151545
#>  [110,]  0.605639103
#>  [111,]  1.755113324
#>  [112,] -0.109400761
#>  [113,]  0.311790577
#>  [114,]  0.500793611
#>  [115,] -0.718701364
#>  [116,] -0.881798212
#>  [117,] -0.133620294
#>  [118,] -0.273921403
#>  [119,] -2.888130973
#>  [120,] -2.302948212
#>  [121,] -2.927965073
#>  [122,] -4.304098774
#>  [123,] -4.643304149
#>  [124,] -4.508437359
#>  [125,] -4.624094147
#>  [126,] -2.185521368
#>  [127,] -2.675410575
#>  [128,] -2.101396081
#>  [129,] -1.518758446
#>  [130,] -2.315541212
#>  [131,] -1.698679600
#>  [132,] -1.501384079
#>  [133,] -2.177447384
#>  [134,] -1.761439246
#>  [135,] -1.457614565
#>  [136,] -3.094097337
#>  [137,] -2.453025684
#>  [138,] -1.205160481
#>  [139,]  0.145916886
#>  [140,]  1.973319282
#>  [141,]  2.921521808
#>  [142,]  2.788629592
#>  [143,]  1.824213304
#>  [144,]  3.087456554
#>  [145,]  4.086898195
#>  [146,]  4.367286093
#>  [147,]  4.675496901
#>  [148,]  4.423354464
#>  [149,]  3.670839195
#>  [150,]  2.581950911
#>  [151,]  2.806561863
#>  [152,]  2.002567679
#>  [153,]  2.350262968
#>  [154,]  1.281876695
#>  [155,]  3.351229138
#>  [156,]  3.956520684
#>  [157,]  1.980201359
#>  [158,]  1.531777952
#>  [159,]  0.438389443
#>  [160,] -0.362423417
#>  [161,] -1.184881857
#>  [162,]  0.046665203
#>  [163,] -0.663661606
#>  [164,] -0.187402242
#>  [165,]  0.285910001
#>  [166,]  1.302437492
#>  [167,]  1.153602411
#>  [168,]  0.402792624
#>  [169,] -0.394836820
#>  [170,] -0.535345401
#>  [171,] -1.270113439
#>  [172,] -1.826340261
#>  [173,] -0.344736905
#>  [174,] -0.849355485
#>  [175,] -0.086447478
#>  [176,]  0.014257405
#>  [177,] -0.427964079
#>  [178,] -0.864869156
#>  [179,] -0.143140032
#>  [180,]  0.066871523
#>  [181,] -0.199660488
#>  [182,] -1.328695106
#>  [183,] -2.813286413
#>  [184,] -3.735020266
#>  [185,] -4.609799186
#>  [186,] -3.207199114
#>  [187,] -3.147028265
#>  [188,] -4.178317541
#>  [189,] -3.575246619
#>  [190,] -4.315726708
#>  [191,] -3.843308603
#>  [192,] -3.645856482
#>  [193,] -4.364016364
#>  [194,] -4.381610405
#>  [195,] -2.886528130
#>  [196,] -1.242460182
#>  [197,] -1.560933613
#>  [198,] -0.503017373
#>  [199,] -0.673054370
#>  [200,]  0.305632213
#>  [201,]  1.075703407
#>  [202,]  0.564890449
#>  [203,] -0.387017869
#>  [204,] -1.852331944
#>  [205,] -1.274066004
#>  [206,] -1.898434377
#>  [207,] -1.482730929
#>  [208,] -0.618866398
#>  [209,] -1.187720831
#>  [210,] -1.686347657
#>  [211,] -1.394073267
#>  [212,] -0.098570051
#>  [213,]  2.208540687
#>  [214,]  4.199244294
#>  [215,]  4.195028811
#>  [216,]  4.691249943
#>  [217,]  6.915367019
#>  [218,]  5.281586452
#>  [219,]  5.589244747
#>  [220,]  4.279283000
#>  [221,]  3.919106416
#>  [222,]  4.159990700
#>  [223,]  4.290746904
#>  [224,]  3.955809028
#>  [225,]  4.644673071
#>  [226,]  3.662335842
#>  [227,]  3.895506833
#>  [228,]  3.853480927
#>  [229,]  2.564886337
#>  [230,]  1.308690044
#>  [231,]  0.547661598
#>  [232,]  2.337920649
#>  [233,]  0.364660323
#>  [234,]  3.451429147
#>  [235,]  2.724556062
#>  [236,]  1.722214051
#>  [237,] -0.529938062
#>  [238,] -0.866152081
#>  [239,] -3.324559989
#>  [240,] -3.184238758
#>  [241,] -3.361588282
#>  [242,] -2.462796339
#>  [243,] -2.950605172
#>  [244,] -2.563106278
#>  [245,] -1.872005044
#>  [246,] -3.741151707
#>  [247,] -2.278441880
#>  [248,] -2.104668053
#>  [249,] -0.679307874
#>  [250,]  0.361900152
#>  [251,]  0.618562176
#>  [252,]  0.826098633
#>  [253,]  0.626068713
#>  [254,]  1.153861804
#>  [255,]  1.928401572
#>  [256,]  1.555812379
#>  [257,]  1.409383244
#>  [258,]  1.594434305
#>  [259,]  2.486785531
#>  [260,]  3.428479583
#>  [261,]  3.356775112
#>  [262,]  4.429767193
#>  [263,]  4.915593358
#>  [264,]  4.062861250
#>  [265,]  3.671081033
#>  [266,]  2.882729663
#>  [267,]  3.971540800
#>  [268,]  3.053481596
#>  [269,]  2.001751146
#>  [270,]  1.033951419
#>  [271,] -1.101263603
#>  [272,]  0.106150186
#>  [273,]  0.045656242
#>  [274,] -1.278629952
#>  [275,] -1.081674350
#>  [276,] -1.443026188
#>  [277,] -0.578495592
#>  [278,] -1.803806938
#>  [279,] -1.216961472
#>  [280,] -1.197966124
#>  [281,] -0.979015535
#>  [282,]  0.703754944
#>  [283,]  0.762105383
#>  [284,]  1.929428456
#>  [285,]  2.980870867
#>  [286,]  3.454486149
#>  [287,]  1.970493241
#>  [288,]  3.225259278
#>  [289,]  3.644963195
#>  [290,]  1.962812407
#>  [291,]  0.904946046
#>  [292,]  1.316122296
#>  [293,]  0.138312653
#>  [294,] -0.223206535
#>  [295,] -0.305721286
#>  [296,] -0.288453427
#>  [297,] -1.234418792
#>  [298,] -0.840618106
#>  [299,] -0.546396936
#>  [300,]  0.648617652
#>  [301,]  1.180711037
#>  [302,]  1.465994654
#>  [303,]  0.395577309
#>  [304,] -0.489481516
#>  [305,] -0.503501866
#>  [306,] -0.561278155
#>  [307,] -0.409140617
#>  [308,] -1.259796824
#>  [309,] -0.261142280
#>  [310,]  1.083173387
#>  [311,]  1.259896533
#>  [312,]  0.304214721
#>  [313,]  2.245327450
#>  [314,]  3.385210775
#>  [315,]  4.914728036
#>  [316,]  2.622627158
#>  [317,]  3.561985932
#>  [318,]  2.873570139
#>  [319,]  3.721380475
#>  [320,]  3.650811896
#>  [321,]  1.083072317
#>  [322,]  0.081777096
#>  [323,] -0.298719606
#>  [324,]  1.167488845
#>  [325,]  1.691076925
#>  [326,]  3.163977144
#>  [327,]  3.626817176
#>  [328,]  3.652656153
#>  [329,]  4.864817980
#>  [330,]  3.817550521
#>  [331,]  3.272633115
#>  [332,]  2.032678146
#>  [333,]  2.710109648
#>  [334,]  5.106504303
#>  [335,]  3.596263550
#>  [336,]  3.897750265
#>  [337,]  2.812706903
#>  [338,]  3.627480599
#>  [339,]  2.063141797
#>  [340,]  2.112537557
#>  [341,]  3.137487903
#>  [342,]  3.176984689
#>  [343,]  2.269186156
#>  [344,]  2.605749816
#>  [345,]  1.978087417
#>  [346,]  1.238138944
#>  [347,]  0.469813534
#>  [348,] -2.367058904
#>  [349,] -3.291680946
#>  [350,] -2.118173553
#>  [351,] -3.293529937
#>  [352,] -4.871807585
#>  [353,] -4.698341305
#>  [354,] -3.295453991
#>  [355,] -1.648699179
#>  [356,] -3.797984566
#>  [357,] -3.003003065
#>  [358,] -2.397319616
#>  [359,] -3.499798724
#>  [360,] -3.470240001
#>  [361,] -2.429693309
#>  [362,] -3.030777256
#>  [363,] -2.292803987
#>  [364,] -3.167448953
#>  [365,] -0.885883516
#>  [366,] -2.697528263
#>  [367,] -2.457587145
#>  [368,] -5.142373906
#>  [369,] -4.115263555
#>  [370,] -3.578156668
#>  [371,] -3.437310735
#>  [372,] -1.767847965
#>  [373,]  0.160531939
#>  [374,] -1.149437226
#>  [375,] -1.176299185
#>  [376,] -2.029498804
#>  [377,] -2.080705104
#>  [378,] -2.984060323
#>  [379,] -4.154155403
#>  [380,] -4.611856553
#>  [381,] -6.076407534
#>  [382,] -6.554369679
#>  [383,] -6.718791388
#>  [384,] -5.870141511
#>  [385,] -4.926884466
#>  [386,] -3.653051470
#>  [387,] -4.597835317
#>  [388,] -3.503761757
#>  [389,] -0.938446248
#>  [390,] -0.867635372
#>  [391,] -0.191215060
#>  [392,] -0.037929702
#>  [393,]  1.316968898
#>  [394,]  2.593152863
#>  [395,]  2.707177037
#>  [396,]  3.133289496
#>  [397,]  3.840494057
#>  [398,]  2.515569730
#>  [399,]  1.071098845
#>  [400,]  1.604756327
#>  [401,] -0.181979545
#>  [402,] -0.103460333
#>  [403,] -0.265135659
#>  [404,]  0.054315758
#>  [405,]  1.257479255
#>  [406,]  0.743987069
#>  [407,]  0.859594422
#>  [408,]  0.842001494
#>  [409,]  0.394780509
#>  [410,]  1.235688514
#>  [411,] -0.314543116
#>  [412,] -0.700215659
#>  [413,] -1.617465724
#>  [414,] -3.599843800
#>  [415,] -1.098697594
#>  [416,] -0.837388657
#>  [417,] -1.912702562
#>  [418,] -2.354868849
#>  [419,] -3.362841910
#>  [420,] -4.002970692
#>  [421,] -1.598008133
#>  [422,] -2.450362242
#>  [423,] -0.681075076
#>  [424,] -0.758648784
#>  [425,] -0.040112777
#>  [426,]  0.821281309
#>  [427,]  0.709317655
#>  [428,]  0.926966847
#>  [429,]  0.957258349
#>  [430,]  1.347914741
#>  [431,]  0.711123811
#>  [432,]  1.090968519
#>  [433,]  1.520747520
#>  [434,]  0.767285851
#>  [435,]  0.188131102
#>  [436,]  0.749776226
#>  [437,] -0.383478630
#>  [438,] -0.370139151
#>  [439,] -1.757278701
#>  [440,] -0.797356660
#>  [441,] -1.053944540
#>  [442,] -0.845890575
#>  [443,]  0.282789051
#>  [444,] -2.740064271
#>  [445,] -2.972604791
#>  [446,] -2.136674249
#>  [447,] -3.206149538
#>  [448,] -2.795728872
#>  [449,] -3.231287293
#>  [450,] -3.290267551
#>  [451,] -2.632844796
#>  [452,] -4.731716864
#>  [453,] -3.360285150
#>  [454,] -2.195750374
#>  [455,] -2.256330850
#>  [456,] -0.929541743
#>  [457,] -1.900833014
#>  [458,] -2.910872275
#>  [459,] -1.650106193
#>  [460,] -1.121552985
#>  [461,] -2.229849468
#>  [462,] -3.126128778
#>  [463,] -2.516459424
#>  [464,] -1.129689968
#>  [465,] -0.790697405
#>  [466,]  0.387213345
#>  [467,]  0.321201455
#>  [468,]  1.004245845
#>  [469,]  0.030507997
#>  [470,] -0.570587287
#>  [471,]  0.849423202
#>  [472,]  1.409332144
#>  [473,]  1.803922051
#>  [474,]  1.244747711
#>  [475,]  2.733907341
#>  [476,]  2.565798239
#>  [477,]  1.302086706
#>  [478,]  1.302227742
#>  [479,] -1.598276420
#>  [480,] -1.432774514
#>  [481,] -0.170951624
#>  [482,]  1.004686142
#>  [483,]  1.242145160
#>  [484,] -0.643419217
#>  [485,] -0.985959232
#>  [486,] -0.168128109
#>  [487,] -0.330575300
#>  [488,]  0.211566065
#>  [489,] -0.794535003
#>  [490,]  0.058289099
#>  [491,] -0.226439734
#>  [492,]  1.648050450
#>  [493,]  1.329466325
#>  [494,]  0.508979616
#>  [495,] -1.291623953
#>  [496,] -1.621458384
#>  [497,] -1.701949646
#>  [498,] -3.532462667
#>  [499,] -5.009286659
#>  [500,] -2.537299218
#>  [501,] -2.106254978
#>  [502,] -2.627467633
#>  [503,] -2.568451554
#>  [504,] -3.035096204
#>  [505,] -1.976885161
#>  [506,] -1.778307501
#>  [507,] -1.621536035
#>  [508,] -1.987103606
#>  [509,] -2.366488331
#>  [510,] -1.520378450
#>  [511,] -0.455639653
#>  [512,] -0.285756605
#>  [513,] -1.163953088
#>  [514,] -1.167205871
#>  [515,] -0.242493082
#>  [516,] -0.170879768
#>  [517,] -0.640129995
#>  [518,] -1.194049741
#>  [519,] -1.255878218
#>  [520,] -0.333565517
#>  [521,] -1.099962510
#>  [522,] -1.111205982
#>  [523,] -1.991190044
#>  [524,] -1.788975556
#>  [525,] -0.766626367
#>  [526,]  1.100688432
#>  [527,]  1.873936207
#>  [528,]  2.815348728
#>  [529,]  2.849449519
#>  [530,]  2.022637749
#>  [531,]  1.814830750
#>  [532,]  1.360237453
#>  [533,]  1.481800834
#>  [534,]  1.008468154
#>  [535,] -0.490327798
#>  [536,] -2.335650838
#>  [537,] -0.445966713
#>  [538,] -3.208084584
#>  [539,] -1.773849505
#>  [540,] -0.841744987
#>  [541,]  0.563938070
#>  [542,]  0.226943540
#>  [543,]  0.478353556
#>  [544,]  1.365675463
#>  [545,]  3.120043896
#>  [546,]  3.317752290
#>  [547,]  1.694790531
#>  [548,]  0.862254276
#>  [549,]  0.706517222
#>  [550,] -2.185249498
#>  [551,] -2.282245882
#>  [552,] -2.704099756
#>  [553,] -2.128666983
#>  [554,] -2.319096797
#>  [555,] -1.569081452
#>  [556,] -3.279745716
#>  [557,] -4.406927230
#>  [558,] -5.642229004
#>  [559,] -4.804960133
#>  [560,] -4.261404060
#>  [561,] -5.118548252
#>  [562,] -4.700028315
#>  [563,] -4.854729014
#>  [564,] -1.887726619
#>  [565,] -2.398554891
#>  [566,] -2.400988958
#>  [567,]  0.180331495
#>  [568,] -0.202386925
#>  [569,]  0.337322072
#>  [570,] -0.165929909
#>  [571,]  2.477190286
#>  [572,]  1.924207245
#>  [573,]  2.652049122
#>  [574,]  2.576607563
#>  [575,]  2.859356663
#>  [576,]  4.993508875
#>  [577,]  6.286074068
#>  [578,]  7.599604019
#>  [579,]  6.699710703
#>  [580,]  4.966421961
#>  [581,]  6.062523447
#>  [582,]  6.785846295
#>  [583,]  5.958339782
#>  [584,]  4.300145505
#>  [585,]  4.779260872
#>  [586,]  4.178232595
#>  [587,]  3.335384648
#>  [588,]  2.095428609
#>  [589,]  2.092823246
#>  [590,]  2.564957265
#>  [591,]  2.691523947
#>  [592,]  2.591626520
#>  [593,]  2.486882960
#>  [594,]  1.385652875
#>  [595,]  1.031890591
#>  [596,] -1.060655205
#>  [597,] -1.101623681
#>  [598,] -2.060743579
#>  [599,] -1.654011611
#>  [600,] -0.471953810
#>  [601,]  0.411192009
#>  [602,]  2.611283714
#>  [603,]  1.813608794
#>  [604,]  1.346096961
#>  [605,]  0.411903915
#>  [606,]  0.313158799
#>  [607,]  1.357133238
#>  [608,]  0.960214098
#>  [609,] -0.161093077
#>  [610,] -0.470410958
#>  [611,] -1.673106067
#>  [612,] -1.321348258
#>  [613,] -0.810728824
#>  [614,] -3.530710137
#>  [615,] -3.604229781
#>  [616,] -3.910632375
#>  [617,] -3.717429423
#>  [618,] -4.468445165
#>  [619,] -3.677676841
#>  [620,] -4.446616075
#>  [621,] -3.299902881
#>  [622,] -3.849302914
#>  [623,] -3.909749690
#>  [624,] -2.513100192
#>  [625,] -0.571428803
#>  [626,]  0.998675682
#>  [627,]  0.511540818
#>  [628,]  0.803381254
#>  [629,]  0.255018667
#>  [630,]  0.825079399
#>  [631,]  0.228951874
#>  [632,]  0.028594559
#>  [633,]  1.057528932
#>  [634,]  1.201400754
#>  [635,]  1.158570824
#>  [636,] -0.238336146
#>  [637,]  0.330519468
#>  [638,]  0.521443085
#>  [639,]  1.093209287
#>  [640,]  0.179926779
#>  [641,]  1.374045282
#>  [642,]  2.452636040
#>  [643,]  3.384719089
#>  [644,]  4.570571311
#>  [645,]  3.781588891
#>  [646,]  3.473167087
#>  [647,]  3.060550064
#>  [648,]  5.137829734
#>  [649,]  3.252802009
#>  [650,]  4.322194639
#>  [651,]  2.152398111
#>  [652,] -0.739046041
#>  [653,]  2.550945572
#>  [654,]  2.205653361
#>  [655,]  1.324050651
#>  [656,]  1.600197943
#>  [657,]  0.679078600
#>  [658,]  0.164923294
#>  [659,] -1.354803550
#>  [660,] -3.320499632
#>  [661,]  0.691959415
#>  [662,] -0.038875029
#>  [663,]  0.639770238
#>  [664,]  0.538990427
#>  [665,]  0.373538546
#>  [666,]  0.282247331
#>  [667,]  0.868352066
#>  [668,]  1.207919081
#>  [669,]  0.464646322
#>  [670,] -0.100699156
#>  [671,]  0.267665052
#>  [672,] -0.638014825
#>  [673,]  1.482149572
#>  [674,]  1.791853193
#>  [675,]  3.241447008
#>  [676,]  2.919365188
#>  [677,]  2.715674406
#>  [678,]  2.997190449
#>  [679,]  3.565497956
#>  [680,]  2.619917944
#>  [681,]  2.413609032
#>  [682,]  2.920109562
#>  [683,]  2.823737210
#>  [684,]  5.092176070
#>  [685,]  4.132411010
#>  [686,]  3.508476267
#>  [687,]  0.972772766
#>  [688,]  0.731896449
#>  [689,] -0.114066041
#>  [690,]  1.553324133
#>  [691,]  1.134685373
#>  [692,] -0.805074650
#>  [693,] -1.747662363
#>  [694,] -1.133217087
#>  [695,] -2.003320908
#>  [696,] -1.751389298
#>  [697,] -4.674206566
#>  [698,] -2.992371504
#>  [699,] -3.005808682
#>  [700,] -0.507845623
#>  [701,] -0.139899605
#>  [702,]  0.522627886
#>  [703,] -1.283291809
#>  [704,] -0.755138920
#>  [705,] -1.199564153
#>  [706,] -0.853825617
#>  [707,] -0.730288259
#>  [708,]  0.523918052
#>  [709,] -0.316863914
#>  [710,] -0.325590064
#>  [711,] -0.141544910
#>  [712,] -1.148260075
#>  [713,] -1.278635674
#>  [714,] -0.929464469
#>  [715,] -3.315404379
#>  [716,] -2.503179173
#>  [717,] -3.014776010
#>  [718,] -1.354347286
#>  [719,] -2.894417495
#>  [720,] -1.782985248
#>  [721,] -0.627118110
#>  [722,] -1.490526725
#>  [723,] -0.812229950
#>  [724,] -0.034887831
#>  [725,] -0.134930890
#>  [726,] -0.847214579
#>  [727,] -0.739714600
#>  [728,] -1.200492028
#>  [729,] -2.852611857
#>  [730,] -3.033672711
#>  [731,] -0.678723467
#>  [732,] -0.133627581
#>  [733,] -1.146369615
#>  [734,] -1.294642780
#>  [735,] -2.071729538
#>  [736,] -2.139852069
#>  [737,] -1.320664849
#>  [738,]  0.620023554
#>  [739,]  0.677930120
#>  [740,]  0.330424333
#>  [741,]  0.982297204
#>  [742,]  2.179977298
#>  [743,]  2.648143944
#>  [744,]  1.528508616
#>  [745,]  0.030965477
#>  [746,]  0.178008631
#>  [747,]  0.498186355
#>  [748,] -0.452265313
#>  [749,] -0.726864907
#>  [750,] -1.557925549
#>  [751,] -1.387994576
#>  [752,] -0.516499275
#>  [753,] -1.963625716
#>  [754,] -3.511349223
#>  [755,] -4.600255784
#>  [756,] -5.048940985
#>  [757,] -5.102960590
#>  [758,] -4.211727962
#>  [759,] -4.316735337
#>  [760,] -5.136999354
#>  [761,] -4.097644296
#>  [762,] -2.896323325
#>  [763,] -1.451382003
#>  [764,] -1.760069154
#>  [765,] -0.386349547
#>  [766,] -1.891499694
#>  [767,] -1.237093513
#>  [768,] -1.437620720
#>  [769,] -0.717698640
#>  [770,]  1.307979713
#>  [771,]  1.184466271
#>  [772,]  0.524644728
#>  [773,]  1.935841225
#>  [774,]  2.295662427
#>  [775,]  2.253374885
#>  [776,]  2.446214232
#>  [777,]  2.749563765
#>  [778,]  1.933001146
#>  [779,]  1.347697444
#>  [780,]  1.517830428
#>  [781,]  2.399607147
#>  [782,]  0.648885453
#>  [783,]  0.505177569
#>  [784,] -1.153051941
#>  [785,] -0.519554715
#>  [786,] -0.842766417
#>  [787,] -0.719788824
#>  [788,] -0.153167175
#>  [789,] -0.600792101
#>  [790,] -1.241858448
#>  [791,] -0.319639753
#>  [792,]  0.613195686
#>  [793,]  0.300993545
#>  [794,] -0.100027259
#>  [795,]  1.338327383
#>  [796,]  1.265446067
#>  [797,]  3.035855628
#>  [798,]  2.395567916
#>  [799,]  0.985631865
#>  [800,]  0.054087056
#>  [801,]  0.794053657
#>  [802,]  0.436160533
#>  [803,]  1.632137638
#>  [804,]  2.347228427
#>  [805,]  0.516666951
#>  [806,]  0.532547488
#>  [807,]  0.189798952
#>  [808,]  0.033754604
#>  [809,] -0.541907855
#>  [810,] -2.060164684
#>  [811,] -2.104098297
#>  [812,] -0.556486845
#>  [813,] -0.729787111
#>  [814,]  0.235305549
#>  [815,] -0.721860519
#>  [816,] -1.506211483
#>  [817,] -1.112588351
#>  [818,] -0.871476252
#>  [819,]  0.312083010
#>  [820,]  1.307019136
#>  [821,]  0.142391410
#>  [822,]  1.677942206
#>  [823,]  3.054495233
#>  [824,]  0.489990611
#>  [825,]  0.303108072
#>  [826,]  0.336497994
#>  [827,]  0.769918663
#>  [828,] -0.166822762
#>  [829,]  1.355474772
#>  [830,]  1.718772792
#>  [831,]  0.300722106
#>  [832,] -0.298927531
#>  [833,] -0.226187323
#>  [834,]  0.184492884
#>  [835,]  2.136107364
#>  [836,]  2.074759747
#>  [837,]  0.235151454
#>  [838,] -0.353308406
#>  [839,] -1.387707618
#>  [840,] -2.734708605
#>  [841,] -1.898583259
#>  [842,] -2.350126351
#>  [843,] -2.451329885
#>  [844,] -3.365792445
#>  [845,] -1.975787529
#>  [846,] -0.702984281
#>  [847,]  0.015032377
#>  [848,] -1.298542742
#>  [849,] -1.456343218
#>  [850,] -2.744797867
#>  [851,] -1.905172152
#>  [852,] -1.590932862
#>  [853,] -1.330878530
#>  [854,] -1.069756060
#>  [855,] -1.216212128
#>  [856,] -0.708026953
#>  [857,] -2.142496801
#>  [858,] -0.169147858
#>  [859,]  0.484738099
#>  [860,] -0.298190713
#>  [861,] -1.536935963
#>  [862,] -1.024354662
#>  [863,] -1.579992716
#>  [864,] -1.802088824
#>  [865,] -2.033690345
#>  [866,] -2.095320410
#>  [867,] -2.427294736
#>  [868,] -2.822989474
#>  [869,] -2.595639571
#>  [870,] -2.228638693
#>  [871,] -2.212279238
#>  [872,] -1.456560597
#>  [873,] -3.259207574
#>  [874,] -3.863502775
#>  [875,] -2.717589527
#>  [876,] -1.924030548
#>  [877,] -2.006093774
#>  [878,] -2.286336440
#>  [879,] -2.880810058
#>  [880,] -3.583473805
#>  [881,] -1.116880714
#>  [882,] -3.311368853
#>  [883,] -4.997582292
#>  [884,] -5.650492388
#>  [885,] -5.298418969
#>  [886,] -4.494700130
#>  [887,] -4.142044249
#>  [888,] -4.323886882
#>  [889,] -3.402407114
#>  [890,] -3.903871942
#>  [891,] -4.730497875
#>  [892,] -4.685272929
#>  [893,] -4.148702038
#>  [894,] -3.376684985
#>  [895,] -2.566560402
#>  [896,] -2.840094232
#>  [897,] -2.965442905
#>  [898,] -2.272211548
#>  [899,] -3.236416701
#>  [900,] -4.760011095
#>  [901,] -4.634645316
#>  [902,] -4.648027072
#>  [903,] -2.443156839
#>  [904,] -1.846752542
#>  [905,] -1.716132251
#>  [906,] -0.684096537
#>  [907,] -0.547344032
#>  [908,] -0.832646962
#>  [909,]  0.747233106
#>  [910,]  0.905273426
#>  [911,]  1.250419480
#>  [912,]  3.027375959
#>  [913,]  2.942574175
#>  [914,]  2.148458170
#>  [915,]  3.012102909
#>  [916,]  1.651450098
#>  [917,]  1.420200984
#>  [918,]  3.061079049
#>  [919,]  1.046484130
#>  [920,]  0.995988969
#>  [921,]  0.849843408
#>  [922,]  0.933181737
#>  [923,]  0.283546623
#>  [924,] -0.177735705
#>  [925,] -0.598779127
#>  [926,] -1.384388997
#>  [927,] -0.294855466
#>  [928,]  0.534451597
#>  [929,]  0.128741927
#>  [930,]  2.022313558
#>  [931,]  0.275488403
#>  [932,]  0.349269124
#>  [933,]  3.189218942
#>  [934,]  3.300241069
#>  [935,]  2.463212444
#>  [936,]  2.928877346
#>  [937,]  2.652933911
#>  [938,]  0.492716697
#>  [939,] -1.071456260
#>  [940,] -0.168185998
#>  [941,]  0.548560585
#>  [942,]  0.368743183
#>  [943,] -0.735589542
#>  [944,] -2.826549701
#>  [945,] -2.461375973
#>  [946,] -0.455669954
#>  [947,]  0.441088571
#>  [948,]  1.042545992
#>  [949,]  2.265326165
#>  [950,]  2.096674956
#>  [951,]  3.212733606
#>  [952,]  2.705386107
#>  [953,]  1.740844558
#>  [954,]  1.622185629
#>  [955,]  2.247391936
#>  [956,]  3.530885023
#>  [957,]  2.800807345
#>  [958,]  1.673937523
#>  [959,]  1.725890396
#>  [960,]  1.068617924
#>  [961,] -0.069185544
#>  [962,] -0.495734906
#>  [963,] -0.192179585
#>  [964,] -1.000821431
#>  [965,] -1.943229070
#>  [966,] -2.834679317
#>  [967,] -2.191158177
#>  [968,] -2.780874838
#>  [969,] -3.141993545
#>  [970,] -3.037066021
#>  [971,] -1.194623629
#>  [972,]  0.001697213
#>  [973,]  0.553083744
#>  [974,]  1.484820729
#>  [975,]  1.222364024
#>  [976,]  2.796745315
#>  [977,]  2.532198480
#>  [978,]  1.032895261
#>  [979,]  0.391365773
#>  [980,]  0.578982610
#>  [981,] -1.312212526
#>  [982,] -1.907892276
#>  [983,] -2.558410070
#>  [984,] -3.429876533
#>  [985,] -2.667335939
#>  [986,] -2.166954736
#>  [987,] -1.367626522
#>  [988,] -2.673284548
#>  [989,] -3.529115643
#>  [990,] -3.771561076
#>  [991,] -2.008737873
#>  [992,] -3.962622381
#>  [993,] -4.954526106
#>  [994,] -4.476860639
#>  [995,] -5.323377220
#>  [996,] -4.116402663
#>  [997,] -2.151862897
#>  [998,] -4.730976752
#>  [999,] -4.982454543
#> [1000,] -5.649904719