Generate Seasonal Autoregressive Order P - Moving Average Order Q (SARMA(p,q)x(P,Q)) Model

Generate an ARMA(P,Q) process with supplied vector of Autoregressive Coefficients (\(\phi\)), Moving Average Coefficients (\(\theta\)), and \(\sigma^2\).

gen_sarima(N, ar, d, ma, sar, sd, sma, sigma2 = 1.5, s = 12L,
  n_start = 0L)

Arguments

N
An integer for signal length.

ar
A vec that contains the AR coefficients.

d
An integer that indicates a non-seasonal difference.

ma
A vec that contains the MA coefficients.

sar
A vec that contains the SAR coefficients.

sd
An integer that indicates a seasonal difference.

sma
A vec that contains the SMA coefficients.

sigma2
A double that contains process variance.

s
An integer that contains a seasonal id.

n_start
An unsigned int that indicates the amount of observations to be used for the burn in period.

Value

A vec that contains the generated observations.

Details

The innovations are generated from a normal distribution. The \(\sigma^2\) parameter is indeed a variance parameter. This differs from R's use of the standard deviation, \(\sigma\).

Examples

gen_sarima(10, c(.3,.5), 1, c(.1), c(.2), 0, c(.4), 1, 12, 0)

#>            [,1]
#>  [1,]  1.671356
#>  [2,]  2.260205
#>  [3,]  4.329881
#>  [4,]  5.462063
#>  [5,]  7.901788
#>  [6,] 10.178695
#>  [7,] 12.013144
#>  [8,] 11.831188
#>  [9,] 11.079839
#> [10,]  9.655307