Generate a Gaussian White Noise Process (WN(\(\sigma ^2\)))

Simulates a Gaussian White Noise Process with variance parameter \(\sigma ^2\).

gen_wn(N, sigma2 = 1)

Arguments

N
An integer for signal length.

sigma2
A double that contains process variance.

Value

wn A vec containing the white noise.

Process Definition

Gaussian White Noise (WN) with parameter \(\sigma^2 \in {\rm I\!R}^{+}\). This process is defined as \(X_t\sim N(0,\sigma^2)\) and is sometimes referred to as Angle (Velocity) Random Walk.

Generation Algorithm

To generate the Gaussian White Noise (WN) process, we first obtain the standard deviation from the variance by taking a square root. Then, we sample \(N\) times from a \(N(0,\sigma ^2)\) distribution.

Examples

gen_wn(10, 1.5)

#>              [,1]
#>  [1,]  1.04968484
#>  [2,] -0.36179637
#>  [3,] -0.05427985
#>  [4,]  1.13915618
#>  [5,] -2.08087560
#>  [6,]  2.18751955
#>  [7,]  0.77669765
#>  [8,] -2.87533519
#>  [9,] -2.02341610
#> [10,]  0.67412424