Get N Colors

Creates n colors from specific palette

getColors(palette, n, rm = NULL)

Arguments

palette
A string that indicates the name of the palette.

n
An integer that specifies the number of colors to be generated.

rm
An integer vector that specifies the index of colors to be removed.

Details

Currently, only the palette 'Set1' is implemented in this function.

'Set1' palette in the package RColorBrewer originally has 9 colors, but we delete the yellow color (the 6th one) since it is too much vibrant.

If n is larger than the mamixum number of colors in the palette, the function will repeat the palette until n colors is created.

Examples

getColors('Set1', 10)

#>  [1] "#E41A1C" "#377EB8" "#4DAF4A" "#984EA3" "#FF7F00" "#A65628" "#F781BF"
#>  [8] "#999999" "#E41A1C" "#377EB8"