Engine for obtaining the GMWM Estimator

This function uses the Generalized Method of Wavelet Moments (GMWM) to estimate the parameters of a time series model.

gmwm_engine(
  theta,
  desc,
  objdesc,
  model_type,
  wv_empir,
  omega,
  scales,
  starting
)

Arguments

theta

A vec with dimensions N x 1 that contains user-supplied initial values for parameters

desc

A vector<string> indicating the models that should be considered.

objdesc

A field<vec> containing a list of parameters (e.g. AR(1) = c(1,1), ARMA(p,q) = c(p,q,1))

model_type

A string that represents the model transformation

wv_empir

A vec that contains the empirical wavelet variance

omega

A mat that represents the covariance matrix.

scales

A vec that contains the scales or taus (2^(1:J))

starting

A bool that indicates whether we guessed starting (T) or the user supplied estimates (F).

Value

A vec that contains the parameter estimates from GMWM estimator.

Details

If type = "imu" or "ssm", then parameter vector should indicate the characters of the models that compose the latent or state-space model. The model options are:

"AR1"a first order autoregressive process with parameters \((\phi,\sigma^2)\)

"ARMA"an autoregressive moving average process with parameters \((\phi _p, \theta _q, \sigma^2)\)

"DR"a drift with parameter \(\omega\)

"QN"a quantization noise process with parameter \(Q\)

"RW"a random walk process with parameter \(\sigma^2\)

"WN"a white noise process with parameter \(\sigma^2\)

If model_type = "imu" or type = "ssm" then starting values pass through an initial bootstrap and pseudo-optimization before being passed to the GMWM optimization. If robust = TRUE the function takes the robust estimate of the wavelet variance to be used in the GMWM estimation procedure.

References

Wavelet variance based estimation for composite stochastic processes, S. Guerrier and Robust Inference for Time Series Models: a Wavelet-Based Framework, S. Guerrier

Author

JJB