GMWM for (Robust) Inertial Measurement Units (IMUs)

Performs the GMWM estimation procedure using a parameter transform and sampling scheme specific to IMUs.

gmwm_imu(model, data, compute.v = "fast", robust = F, eff = 0.6, ...)

Arguments

model
A ts.model object containing one of the allowed models.

data
A matrix or data.frame object with only column (e.g. \(N \times 1\)), or a lts object, or a gts object.

compute.v
A string indicating the type of covariance matrix solver. "fast", "bootstrap", "asymp.diag", "asymp.comp", "fft"

robust
A boolean indicating whether to use the robust computation (TRUE) or not (FALSE).

eff
A double between 0 and 1 that indicates the efficiency.

...
Other arguments passed to the main gmwm function

Value

A gmwm object with the structure:

estimateEstimated Parameters Values from the GMWM Procedure

init.guessInitial Starting Values given to the Optimization Algorithm

wv.empirThe data's empirical wavelet variance

ci_lowLower Confidence Interval

ci_highUpper Confidence Interval

orgVOriginal V matrix

VUpdated V matrix (if bootstrapped)

omegaThe V matrix inversed

obj.funValue of the objective function at Estimated Parameter Values

theoSummed Theoretical Wavelet Variance

decomp.theoDecomposed Theoretical Wavelet Variance by Process

scalesScales of the GMWM Object

robustIndicates if parameter estimation was done under robust or classical

effLevel of efficiency of robust estimation

model.typeModels being guessed

compute.vType of V matrix computation

augmentedIndicates moments have been augmented

alphaAlpha level used to generate confidence intervals

expect.diffMean of the First Difference of the Signal

NLength of the Signal

GNumber of Guesses Performed

HNumber of Bootstrap replications

KNumber of V matrix bootstraps

modelts.model supplied to gmwm

model.hatA new value of ts.model object supplied to gmwm

startingIndicates whether the procedure used the initial guessing approach

seedRandomization seed used to generate the guessing values

freqFrequency of data

Details

This version of the gmwm function has customized settings ideal for modeling with an IMU object. If you seek to model with an Gauss Markov, GM, object. Please note results depend on the freq specified in the data construction step within the imu. If you wish for results to be stable but lose the ability to interpret with respect to freq, then use AR1 terms.