Create a GMWM TS Object based on data

Setups a time series oriented object that works well with graphing and summary utilities

gts(data, start = 0, end = NULL, freq = 1, unit = NULL, name = NULL)

Arguments

data
A one-column matrix, data.frame, or a numeric vector.

start
A numeric that provides the time of the first observation.

end
A numeric that provides the time of the last observation.

freq
A numeric that provides the rate of samples. Default value is 1.

unit
A string that contains the unit expression of the frequency. Default value is NULL.

name
A string that provides an identifier to the data. Default value is NULL.

Value

A gts object with the following attributes:

start

The time of the first observation

end

The time of the last observation

freq

Numeric representation of frequency

unit

String representation of the unit

name

Name of the dataset

Examples

m = data.frame(rnorm(50))
x = gts(m, unit = 'sec', name = 'example')
x

#>                example
#>  1: -0.156666894649192
#>  2:  -2.50383733966227
#>  3:  -1.08495472224127
#>  4:  0.669820623021991
#>  5:   1.30495036447416
#>  6:   -1.5295112051031
#>  7:   1.37485234057825
#>  8:  -1.34116148638815
#>  9:  -1.30285133163884
#> 10:  0.257888920956716
#> ---                   
#> 41:   1.20040237647258
#> 42:  -1.10558209976912
#> 43: -0.948833396923832
#> 44:  0.291815555154494
#> 45: -0.158027688120339
#> 46:  0.233379712424301
#> 47: -0.297389569270799
#> 48:   1.74988012568865
#> 49: -0.848661427463536
#> 50: -0.635294701859219

x = gen_gts(50, WN(sigma2 = 1))
x = gts(x, freq = 100, unit = 'sec')