Randomly guess a starting parameter

Sets starting parameters for each of the given parameters.

guess_initial(
  desc,
  objdesc,
  model_type,
  num_param,
  expect_diff,
  N,
  wv,
  tau,
  ranged,
  G
)

Arguments

desc
A vector<string> that contains the model's components.

objdesc
A field<vec> that contains an object description (e.g. values) of the model.

model_type
A string that indicates whether it is an SSM or sensor.

num_param
An unsigned int number of parameters in the model (e.g. # of thetas).

expect_diff
A double that contains the mean of the first difference of the data

N
A integer that contains the number of observations in the data.

tau
A vec that contains the scales. (e.g. 2^(1:J))

G
A integer that indicates how many random draws that should be performed.

wv_empir
A vec that contains the empirical wavelet variance.

double
A double that contains the drift slope given by \(\frac{max-min}{N}\)

Value

A vec containing smart parameter starting guesses to be iterated over.