Generate the ts model object's theta vector

Creates the ts.model's theta vector

model_theta(desc)

Arguments

desc
A vector<string> that contains a list of the strings of each process.

Value

A vec with values initialized at 0 that span the space of parameters to be estimated.

Details

This function currently does NOT support ARMA(P,Q) models. That is, there is no support for ARMA(P,Q), AR(P), or MA(Q). There is support for ARMA11, AR1, MA1, GM, WN, DR, QN, and RW.