Computes the (MODWT) wavelet variance

Calculates the (MODWT) wavelet variance

modwt_wvar_cpp(
  signal,
  nlevels,
  robust,
  eff,
  alpha,
  ci_type,
  strWavelet,
  decomp
)

Arguments

signal
A vec that contains the data.

robust
A boolean that triggers the use of the robust estimate.

eff
A double that indicates the efficiency as it relates to an MLE.

alpha
A double that indicates the \(\left(1-p\right)\times \alpha\) confidence level

ci_type
A string indicating the confidence interval being calculated. Valid value: "eta3"

strWavelet
A string indicating the type of wave filter to be applied. Must be "haar"

decomp
A string indicating whether to use "modwt" or "dwt" decomp

Value

A mat with the structure:

"variance"Wavelet Variance

"low"Lower CI

"high"Upper CI

Details

This function powers the wvar object. It is also extendable...