Multiply a ts.model by constant

Sets up the necessary backend for creating multiple model objects.

# S3 method for ts.model
print(x, ...)

Arguments

x

A numeric value

...

further arguments passed to or from other methods.

Value

An S3 object with called ts.model with the following structure:

desc

theta

Author

James Balamuta

Examples

# Creates a parameter space for guessing
QN() + DR() + WN() + RW() + AR1() + ARMA(1,2)
#> 
#> Guess Starting Values: TRUE 
#> The program will attempt to guess starting values for...
#>  [1] "QN"     "DR"     "WN"     "RW"     "AR1"    "SIGMA2" "AR"     "MA"    
#>  [9] "MA"     "SIGMA2"
#> To have the option of using your own starting values, please supply values for each parameter.

# Creates a user-specified starting value model
AR1(phi = .9, sigma2 = .1) + WN(sigma2 = 1) 
#> 
#> Guess Starting Values: FALSE 
#>   Terms Initial.Values
#>     AR1            0.9
#>  SIGMA2            0.1
#>      WN            1.0
#> The model will be initiated using the initial values you supplied.

# Similarly, with the addition of a generic ARMA
RW(gamma2 = .3) + DR(omega = .5) + QN(q2 = .9) + ARMA(ar = c(.3,.1), ma = c(.3,.2), sigma2 = .99)
#> 
#> Guess Starting Values: FALSE 
#>   Terms Initial.Values
#>      RW           0.30
#>      DR           0.50
#>      QN           0.90
#>      AR           0.30
#>      AR           0.10
#>      MA           0.30
#>      MA           0.20
#>  SIGMA2           0.99
#> The model will be initiated using the initial values you supplied.

# In a similar vein, this example highlights the lack of need for specifying parameters. 
AR1(.9,.1) + WN(1) + RW(.3) + DR(.5) + QN(.9) + ARMA(c(.3,.1), c(.3,.2), .99)
#> 
#> Guess Starting Values: FALSE 
#>   Terms Initial.Values
#>     AR1           0.90
#>  SIGMA2           0.10
#>      WN           1.00
#>      RW           0.30
#>      DR           0.50
#>      QN           0.90
#>      AR           0.30
#>      AR           0.10
#>      MA           0.30
#>      MA           0.20
#>  SIGMA2           0.99
#> The model will be initiated using the initial values you supplied.