Read an IMU Binary File into R

Process binary files within the

read.imu(file, type, unit = NULL, name = NULL)

Arguments

file
A string containing file names or paths.

type
A string that contains a supported IMU type given below.

unit
A string that contains the unit expression of the frequency. Default value is NULL.

name
A string that provides an identifier to the data. Default value is NULL.

Value

An imu object that contains 3 gyroscopes and 3 accelerometers in that order.

Details

Currently supports the following IMUs:

IMAR

LN200

LN200IG

IXSEA

NAVCHIP_INT

NAVCHIP_FLT

References

Thanks goes to Philipp Clausen of Labo TOPO, EPFL, Switzerland, topo.epfl.ch, Tel:+41(0)21 693 27 55 for providing a matlab function that reads in IMUs. This function is a heavily modified port of MATLAB code into Armadillo/C++.

Examples

if (FALSE) {
# Relative
setwd("F:/")

a = read.imu(file = "Documents/James/short_test_data.imu", type = "IXSEA")

# Fixed path
b = read.imu(file = "F:/Desktop/short_test_data.imu", type = "IXSEA")
}