Read an IMU Binary File into R

The function will take a file location in addition to the type of sensor it came from and read the data into R.

read_imu(file_path, imu_type)

Arguments

file_path
A string that contains the full file path.

imu_type
A string that contains a supported IMU type given below.

Value

A matrix with dimensions N x 7, where the columns represent:

Col 0

Time

Col 1

Gyro 1

Col 2

Gyro 2

Col 3

Gyro 3

Col 4

Accel 1

Col 5

Accel 2

Col 6

Accel 3

Details

Currently supports the following IMUs:

IMAR

LN200

LN200IG

IXSEA

NAVCHIP_INT

NAVCHIP_FLT

We hope to soon be able to support delimited files.

References

Thanks goes to Philipp Clausen of Labo TOPO, EPFL, Switzerland, topo.epfl.ch, Tel:+41(0)21 693 27 55 for providing a matlab function that reads in IMUs. The function below is a heavily modified port of MATLAB code into Armadillo/C++.

Examples

# NOT RUN {
read_imu(file_path = "F:/Desktop/short_test_data.imu", imu_type = "IXSEA")
# }