Time Series Recursive Filters

Applies a recursive filter to a univariate time series.

rfilter(x, filter, init)

Arguments

x
A column vector of length T

filter
A column vector of length f

init
A column vector of length f that contains the initial values of the time series in reverse.

Value

x A column vector with its contents reversed.

Details

Note: The length of 'init' must be equal to the length of 'filter'. This is a port of the rfilter function harnessed by the filter function in stats. It is about 6-7 times faster than R's base function. The benchmark was done on iMac Late 2013 using vecLib as the BLAS.

Examples

x = 1:15
# 
rfilter(x, rep(1, 3), rep(1, 3))

#>        [,1]
#>  [1,]     4
#>  [2,]     8
#>  [3,]    16
#>  [4,]    32
#>  [5,]    61
#>  [6,]   115
#>  [7,]   215
#>  [8,]   399
#>  [9,]   738
#> [10,]  1362
#> [11,]  2510
#> [12,]  4622
#> [13,]  8507
#> [14,] 15653
#> [15,] 28797

# Using R's function
filter(x, rep(1, 3), method="recursive", init=rep(1, 3))

#> Time Series:
#> Start = 1 
#> End = 15 
#> Frequency = 1 
#>  [1]     4     8    16    32    61   115   215   399   738  1362  2510  4622
#> [13]  8507 15653 28797