Create the ts.model obj.desc given split values

Computes the total phi and total theta vector length.

sarma_objdesc(ar, ma, sar, sma, s, i, si)

Arguments

ar
A vec containing the non-seasonal phi parameters.

ma
A vec containing the non-seasonal theta parameters.

sar
A vec containing the seasonal phi parameters.

sma
A vec containing the seasonal theta parameters.

s
An unsigned integer containing the frequency of seasonality.

i
An unsigned integer containing the number of non-seasonal differences.

si
An unsigned integer containing the number of seasonal differences.

Value

A vec with rows:

np

Number of Non-Seasonal AR Terms

nq

Number of Non-Seasonal MA Terms

nsp

Number of Seasonal AR Terms

nsq

Number of Seasonal MA Terms

nsigma

Number of Variances (always 1)

s

Season Value

i

Number of non-seasonal differences

si

Number of Seasonal Differences