Generate a sequence of values

Creates a vector containing a sequence of values starting at the initial point and going to the terminal point.

seq_cpp(a, b)

Arguments

a
An int, that denotes the starting point.

b
An int, that denotes the ending point.

Value

A vector containing values moving from a to b. There are no restrictions on A's range.

Examples

#Call with the following data:
seq_cpp(3, 5)

#>      [,1]
#> [1,]    3
#> [2,]    4
#> [3,]    5

seq_cpp(5, 3)

#>      [,1]
#> [1,]    5
#> [2,]    4
#> [3,]    3