Simplify and print SARIMA model

Simplify and print SARIMA model

simplified_print_SARIMA(p, i, q, P, si, Q, s)

Arguments

p
An integer denoting the length of ar.

i
An integer containing the number of differences to be done.

q
An integer denoting the length of ma.

P
An integer denoting the length of sma.

si
An integer containing the number of seasonal differences to be done.

Q
An integer denoting the length of sar.

s
An integer indicating the seasonal value of the data.

Value

An S3 object with the following structure:

print

String containing simplified model

simplified

Type of model (after simplification)