simts: Time Series Simulation

Generates various time series objects for use in other packages.

Simulation and Plotting Tools Time Series (simts) Package

Details

The data sets in this package may change at a moments notice.

Package:	simts

Type:	Package

Version:	1.0

Date:	2018-09-15

License:	GPL

Package:

See also