Update (Robust) GMWM object for IMU or SSM

Provides a way to estimate different models over the previously estimated wavelet variance values and covariance matrix.

# S3 method for gmwm
update(object, model, ...)

Arguments

object

A gmwm object.

model

A ts.model object containing one of the allowed models

...

Additional parameters (not used)

Value

A gmwm object with the structure:

estimateEstimated Parameters Values from the GMWM Procedure

init.guessInitial Starting Values given to the Optimization Algorithm

wv.empirThe data's empirical wavelet variance

ci_lowLower Confidence Interval

ci_highUpper Confidence Interval

orgVOriginal V matrix

VUpdated V matrix (if bootstrapped)

omegaThe V matrix inversed

obj.funValue of the objective function at Estimated Parameter Values

theoSummed Theoretical Wavelet Variance

decomp.theoDecomposed Theoretical Wavelet Variance by Process

scalesScales of the GMWM Object

robustIndicates if parameter estimation was done under robust or classical

effLevel of efficiency of robust estimation

model.typeModels being guessed

compute.vType of V matrix computation

augmentedIndicates moments have been augmented

alphaAlpha level used to generate confidence intervals

expect.diffMean of the First Difference of the Signal

NLength of the Signal

GNumber of Guesses Performed

HNumber of Bootstrap replications

KNumber of V matrix bootstraps

modelts.model supplied to gmwm

model.hatA new value of ts.model object supplied to gmwm

startingIndicates whether the procedure used the initial guessing approach

seedRandomization seed used to generate the guessing values

freqFrequency of data