Kalman Filter Estimation

Compute the Kalman Filter estimation for the sum of any combinations of AR processes, RW, WN and DR

kalman_filter(model, y, estimate_model = F, model_to_estimate = NULL,
  method = "mle")

Arguments

model
a ts.model object with specified parameters values

y
a lts object

estimate_model
a bolean indicating whether or not to estimate the parameters values of a provided model, default is False

model_to_estimate
a ts.model object without parameters values specified if estimate_model is set to True

method
specify the method of estimation if estimate_model is set to True and model_to_estimate is provided

Value

a KF object with the structure:

forecast

X_t|t-1

forecast_cov_mat

Estimated variance-covariance for X_t|t-1

filter

X_t|t

filter_cov_mat

Estimated variance-covariance for X_t|t

smooth

X_t|t with a forward and backward Kalman filter

smooth_cov_mat

Estimated variance-covariance for the smoothed states

y

Observed time serie

y_d

Unobserved states composing the time serie

print

Model structure

Details

Compute a kalman filter on a state space model which can be composed of the sum of AR(), RW(), DR() and WN() processes. Note that there can be any ammount of AR() processes while RW(), DR() and WN() processes are limited to 1. The function can both estimate the model parameter from a given model and apply the kalman filter or directly apply the kalman filter with specified parameters. The function takes as input the model which is the defined model and y which is the observed time serie. If estimate_model is set to False (by default), the user need to provide a model and its parameters (a ts.model class object). If estimate_model is set to True, the user just need to provide the selected model to estimate and the estimation method (see function estimate).

Note on the calculation of the smoothed states:

The variance-covariance matrix P_s for the smoothed states is calculated as follows:

P_s = (P_f^-1 + P_b^-1)^-1

where P_f and P_b are the variance-covariance matrices of the filtered states for respectively the foward and backward kalman filter algorithm.

The smoothed states X_s are calculated as follows:

X_b + P_s*P_f^-1(X_f - X_b)

where P_f and P_b are the variance-covariance matrices of the filtered states for respectively the foward and backward kalman filter algorithm, and X_f and X_b are the filtered states for respectively the foward and backward kalman filter algorithm.

Examples

#Filter a 2*AR1 + DR + RW + WN process
set.seed(123)
model = AR(.3, 2) + AR(.5,3) + DR(.1) + RW(3) + WN(4)
n = 250
y = gen_lts(n = n, model = model)
my_res = kalman_filter(model = model, y = y)