(Internal) Expand the SARMA Parameters

(Internal) Expand the SARMA Parameters

sarma_expand_unguided(params, np, nq, nsp, nsq, ns, p, q)

Arguments

params
A vec containing the theta values of the parameters.

np
An unsigned int containing the number of non-seasonal phi parameters.

nq
An unsigned int containing the number of non-seasonal theta parameters.

nsp
An unsigned int containing the number of seasonal phi parameters.

nsq
An unsigned int containing the number of seasonal theta parameters.

p
An unsigned int that is the total size of the phi vector.

q
An unsigned int that is the total size of the theta vector.

Value

A field<vec> that contains the expansion.