simts: Time Series Analysis Tools

A system contains easy-to-use tools as a support for time series analysis courses. In particular, it incorporates a technique called Generalized Method of Wavelet Moments (GMWM) as well as its robust implementation for fast and robust parameter estimation of time series models which is described, for example, in Guerrier et al. (2013) <doi: 10.1080/01621459.2013.799920>. More details can also be found in the paper linked to via the URL below.

Details

The data sets in this package may change at a moments notice.

Package:	simts

Type:	Package

Version:	1.0

Date:	2018-09-15

License:	GPL

See also

Author

Maintainer: Stéphane Guerrier stef.guerrier@gmail.com [copyright holder]

Authors:

James Balamuta [copyright holder]

Roberto Molinari [copyright holder]

Justin Lee

Lionel Voirol

Yuming Zhang

Other contributors:

Wenchao Yang [contributor]

Nathanael Claussen [contributor]

Yunxiang Zhang [contributor]

Christian Gunning [copyright holder]

Romain Francois [copyright holder]

Ross Ihaka [copyright holder]

R Core Team [copyright holder]

James Balamuta balamut2@illinois.edu, Stephane Guerrier stef.guerrier@gmail.com, Roberto Molinari roberto.molinari@hotmail.it, Wenchao Yang wyang40@illinois.edu Justin Lee munsheet93@gmail.com Yuming Zhang yuming.zhang@unige.ch