simts: Time Series Simulation

Generates various time series objects for use in other packages.

Simulated Time Series (simts) Package

Details

The data sets in this package may change at a moments notice.

Package: 	 simts

Type: 	 Package

Version: 	 1.0

Date: 	 2017-05-28

License: 	 GPL

See also

Useful links:

https://github.com/SMAC-Group/simts

Report bugs at https://github.com/SMAC-Group/simts/issues